Move logic for rewriting typehints out of Transformer and hide mess behind an interface
