+++ src/PhpSpec/CodeAnalysis/TypeHintRewriter.php
@@ -1,0 +1,24 @@
+<?php
+
+/*
+ * This file is part of PhpSpec, A php toolset to drive emergent
+ * design by specification.
+ *
+ * (c) Marcello Duarte <marcello.duarte@gmail.com>
+ * (c) Konstantin Kudryashov <ever.zet@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace PhpSpec\CodeAnalysis;
+
+interface TypeHintRewriter
+{
+    /**
+     * @param string $classDefinition
+     *
+     * @return string
+     */
+    public function rewrite($classDefinition);
+}
